safari-534.53-branch/LayoutTests/fast/viewport/viewport-77-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key ";" not recognized and ignored.
CONSOLE MESSAGE: line 3: Viewport argument value "yes;" for key "initial-scale" not recognized. Content ignored.
viewport size 1280x1408 scale 0.250000 with limits [0.250000, 5.000000] and userScalable -1.000000
